test(accordion): added back original unspecified test and restructure tests
