+++ packages/core/src/components/accordion/test/basic/primary_darkmode/accordion.e2e.ts-snapshots/tds-accordion-primary-darkmode-renders-basic-a-693c7-rectly-with-primary-mode-variant-and-darkmode-1-linux.png
