+++ packages/core/src/components/accordion/test/basic/primary_darkmode/accordion.e2e.ts-snapshots/tds-accordion-primary-darkmode-second-accordion-item-opens-on-click-1-linux.png
