+++ packages/core/src/components/accordion/test/basic/primary_lightmode/accordion.e2e.ts-snapshots/tds-accordion-primary-lightmode-renders-basic--15d6b-ectly-with-primary-mode-variant-and-lightmode-1-linux.png
