+++ packages/core/src/components/accordion/test/basic/primary_lightmode/accordion.e2e.ts-snapshots/tds-accordion-primary-lightmode-second-accordion-item-opens-on-click-1-linux.png
