+++ packages/core/src/components/accordion/test/basic/secondary_darkmode/accordion.e2e.ts-snapshots/tds-accordion-secondary-darkmode-renders-basic-d2ec3-ctly-with-secondary-mode-variant-and-darkmode-1-linux.png
