+++ packages/core/src/components/accordion/test/basic/secondary_darkmode/accordion.e2e.ts-snapshots/tds-accordion-secondary-darkmode-second-accordion-item-opens-on-click-1-linux.png
