+++ packages/core/src/components/accordion/test/basic/secondary_lightmode/accordion.e2e.ts-snapshots/tds-accordion-secondary-lightmode-renders-basi-f26d5-tly-with-secondary-mode-variant-and-lightmode-1-linux.png
