+++ packages/core/src/components/accordion/test/basic/secondary_lightmode/accordion.e2e.ts-snapshots/tds-accordion-secondary-lightmode-second-accordion-item-opens-on-click-1-linux.png
