+++ packages/core/src/components/accordion/test/basic/unspecified/accordion.e2e.ts
@@ -1,18 +1,13 @@
 import { test } from 'stencil-playwright';
 import { expect } from '@playwright/test';
 
-const componentTestPath = 'src/components/accordion/test/basic_primary/lightmode/index.html';
+const componentTestPath = 'src/components/accordion/test/basic/unspecified/index.html';
 const accordionSelector = 'tds-accordion';
 
-test.describe.parallel('tds-accordion-primary-lightmode', () => {
-  test.beforeEach(async ({ page }) => {
-    await page.goto(componentTestPath);
-  });
-
-  test('renders basic accordion correctly with primary mode variant and lightmode', async ({
-    page,
-  }) => {
+test.describe.parallel('tds-accordion', () => {
+  test('renders basic accordion correctly', async ({ page }) => {
     // Define selector for accordion
+    await page.goto(componentTestPath);
     const accordion = page.locator(accordionSelector);
 
     // Check if accordion contains the correct text
@@ -25,6 +20,7 @@ test.describe.parallel('tds-accordion-primary-lightmode', () => {
 
   test('should handle hover on accordion items', async ({ page }) => {
     // Define selector for first accordion item
+    await page.goto(componentTestPath);
     const accordionFirstItem = page.getByText('First item');
 
     // Hover first accordion item
@@ -46,6 +42,7 @@ test.describe.parallel('tds-accordion-primary-lightmode', () => {
 
   test('second accordion item opens on click', async ({ page }) => {
     // Define selector for second accordion item
+    await page.goto(componentTestPath);
     const accordionSecondItem = page.getByText('Second item');
 
     // Hover second accordion item
@@ -57,6 +54,7 @@ test.describe.parallel('tds-accordion-primary-lightmode', () => {
 
   test('fires tdsToggle event on click', async ({ page }) => {
     // Define selector for first accordion item
+    await page.goto(componentTestPath);
     const accordionFirstItem = page.getByText('First item');
 
     // Define selector for second accordion item
